Consolidate redundant test files into test_100_percent_coverage.py
Merged tests from test_direct_coverage.py and test_server_exception_coverage.py
into test_100_percent_coverage.py to reduce test suite complexity.

Changes:
- Merged TestConversationMemoryServerDirect class (54 tests)
- Merged TestServerExceptionCoverage class (9 tests)
- Removed duplicate test files
- All 416 tests passing with identical coverage

This addresses code review feedback about redundant test files.
