--- tests/test_server_exception_coverage.py
@@ -1,207 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test module for server exception handling coverage in server_fastmcp.py.
-
-This module targets Phase 2 coverage: specific exception paths and error scenarios
-in the ConversationMemoryServer initialization and core methods.
-"""
-
-import json
-import tempfile
-import unittest
-import asyncio
-from pathlib import Path
-from unittest.mock import patch, MagicMock
-import shutil
-
-# Add project root and src directory to path using dynamic resolution
-project_root = Path(__file__).parent.parent
-import sys
-sys.path.insert(0, str(project_root))
-sys.path.insert(0, str(project_root / 'src'))
-
-
-def get_test_temp_dir(prefix="test_"):
-    """Create a test temp directory in the project root (safe for testing)"""
-    return tempfile.mkdtemp(prefix=prefix, dir=str(project_root))
-
-# Mock FastMCP before importing server
-class MockFastMCP:
-    def __init__(self, name=None):
-        self.name = name
-        self.tools = {}
-    
-    def tool(self, name=None, description=None):
-        def decorator(func):
-            self.tools[name or func.__name__] = func
-            return func
-        return decorator
-    
-    def run(self):
-        pass
-
-sys.modules['mcp.server.fastmcp'] = type(sys)('mcp.server.fastmcp')
-sys.modules['mcp.server.fastmcp'].FastMCP = MockFastMCP
-
-from conversation_memory import ConversationMemoryServer
-
-
-class TestServerExceptionCoverage(unittest.TestCase):
-    """Test exception handling in server_fastmcp.py for Phase 2 coverage"""
-    
-    def test_init_exception_handling_lines_96_98(self):
-        """Test __init__ exception handling for SQLite initialization (lines 61-63)"""
-        # Mock SearchDatabase to raise an exception during initialization
-        with patch('conversation_memory.SearchDatabase', 
-                   side_effect=Exception("SQLite initialization failed")):
-            
-            # Create server in project directory to pass security checks
-            temp_dir = get_test_temp_dir("test_sqlite_init_")
-            
-            try:
-                # This should catch the SQLite exception and continue
-                server = ConversationMemoryServer(temp_dir, enable_sqlite=True)
-                # Server should still be created but with SQLite disabled
-                self.assertFalse(server.use_sqlite_search)
-            finally:
-                shutil.rmtree(temp_dir, ignore_errors=True)
-    
-    def test_init_index_files_exception_lines_109_110(self):
-        """Test _init_index_files exception handling (lines 109-110)"""
-        # Create temp dir in system temp directory to avoid project root clutter
-        temp_dir = tempfile.mkdtemp(prefix="test_server_exception_")
-        
-        # Mock json.dump to raise an exception during index file creation
-        with patch('json.dump', side_effect=Exception("JSON write failed")):
-            try:
-                # This should trigger the exception in _init_index_files
-                server = ConversationMemoryServer(temp_dir)
-                # If no exception is raised, the lines might not be covered
-                # But we're testing the path exists
-            except Exception as e:
-                # Exception handling should log and possibly continue
-                self.assertIn("JSON write failed", str(e))
-    
-    def test_index_file_creation_permission_error_lines_115_116(self):
-        """Test index file creation permission errors (lines 115-116)"""
-        temp_dir = tempfile.mkdtemp(prefix="test_permission_")
-        
-        # Mock Path.mkdir to raise PermissionError
-        with patch('pathlib.Path.mkdir', side_effect=PermissionError("Permission denied")):
-            try:
-                server = ConversationMemoryServer(temp_dir)
-                # Test that server handles permission errors gracefully
-            except PermissionError:
-                # Exception might be re-raised or handled
-                pass
-    
-    def test_search_conversations_file_error_lines_212_215(self):
-        """Test search_conversations file reading errors (lines 212-215)"""
-        temp_dir = tempfile.mkdtemp(prefix="test_search_error_")
-        server = ConversationMemoryServer(temp_dir)
-        
-        # Mock file operations to trigger exception handling
-        with patch('builtins.open', side_effect=OSError("File read error")):
-            try:
-                # This should trigger exception handling in search_conversations
-                result = asyncio.run(server.search_conversations("test query"))
-                # Should return error response (could be list with error dict)
-                self.assertTrue(isinstance(result, (str, list)))
-                if isinstance(result, list) and result:
-                    self.assertIn("error", result[0])
-            except OSError:
-                # Exception might propagate or be handled
-                pass
-    
-    def test_add_conversation_file_error_lines_272_274(self):
-        """Test add_conversation file writing errors (lines 272-274)"""
-        temp_dir = tempfile.mkdtemp(prefix="test_add_error_")
-        server = ConversationMemoryServer(temp_dir)
-        
-        # Mock file operations to trigger exception handling  
-        with patch('builtins.open', side_effect=OSError("File write error")):
-            try:
-                # This should trigger exception handling in add_conversation
-                result = asyncio.run(server.add_conversation("test content", "test title", "2025-06-10T12:00:00Z"))
-                # Should return error status or raise exception
-                if isinstance(result, str):
-                    self.assertIn("error", result.lower())
-            except OSError:
-                # Exception might propagate or be handled
-                pass
-    
-    def test_missing_conversation_file_lines_493_494(self):
-        """Test missing conversation file handling (lines 493-494)"""
-        temp_dir = tempfile.mkdtemp(prefix="test_missing_file_")
-        server = ConversationMemoryServer(temp_dir)
-        
-        # Create a mock conversation entry that points to non-existent file
-        mock_conv_info = {
-            "file_path": "2025/06-june/non-existent-file.md",
-            "title": "Test Conversation"
-        }
-        
-        # This should trigger lines 493-494 in _analyze_conversations method
-        # when checking if conversation file exists
-        result = server._analyze_conversations([mock_conv_info])
-        
-        # The method should handle missing files gracefully
-        self.assertIsInstance(result, list)
-    
-    def test_additional_error_handling_lines_507_510(self):
-        """Test additional error handling scenarios (lines 507-510)"""
-        temp_dir = tempfile.mkdtemp(prefix="test_additional_error_")
-        server = ConversationMemoryServer(temp_dir)
-        
-        # Test various error scenarios that might trigger lines 507-510
-        # This could be in file operations, JSON processing, etc.
-        
-        # Create a conversation file that exists but can't be read  
-        conv_dir = Path(temp_dir) / "conversations" / "2025" / "06-june"
-        conv_dir.mkdir(parents=True, exist_ok=True)
-        conv_file = conv_dir / "test-conversation.md"
-        conv_file.write_text("Test content")
-        
-        mock_conv_info = {
-            "file_path": "2025/06-june/test-conversation.md",
-            "title": "Test Conversation"
-        }
-        
-        # Mock file reading to raise an exception (triggers lines 507-510)
-        with patch('builtins.open', side_effect=OSError("File read error")):
-            # This should trigger lines 507-510 exception handling
-            result = server._analyze_conversations([mock_conv_info])
-            
-            # Should handle the error gracefully and return default values
-            self.assertIsInstance(result, list)
-    
-    def test_initialization_with_invalid_permissions(self):
-        """Test server initialization with permission issues"""
-        # Try to create server in a restricted directory (triggers security validation)
-        restricted_path = "/root/restricted_test"  # Should fail due to security validation
-        
-        with self.assertRaises(PermissionError) as context:
-            server = ConversationMemoryServer(restricted_path)
-        
-        # This covers permission error during path detection
-        self.assertIn("Permission denied", str(context.exception))
-    
-    def test_malformed_storage_path_handling(self):
-        """Test handling of malformed storage paths"""
-        malformed_paths = [
-            "",  # Empty path
-            None,  # None path (will be converted to string)
-            "invalid\x00path",  # Path with null character
-        ]
-        
-        for path in malformed_paths:
-            try:
-                server = ConversationMemoryServer(str(path) if path is not None else "None")
-                # Some paths might be accepted and handled
-            except (ValueError, OSError, TypeError) as e:
-                # Expected for malformed paths
-                self.assertIsInstance(e, (ValueError, OSError, TypeError))
-
-
-if __name__ == '__main__':
-    unittest.main()
